CORDA-3377: Make API more Java-friendly and resolve some simple warnings. (#125)

+++ djvm/src/main/java/sandbox/java/lang/Float.java
@@ -56,7 +56,7 @@ public final class Float extends Number implements Comparable<Float> {
 
     @Override
     public double doubleValue() {
-        return (double)value;
+        return value;
     }
 
     @Override
